// src/types/generated/Project.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { TextFormat } from "./TextFormat";

export type Project = { id: string, parentId: string | null, title: string, description: string, descriptionFormat: TextFormat, sortOrder: number, createdAt: string, updatedAt: string, deletedAt: string | null, };